desenvolvendo editar produtos(imagens e categorias)

+++ controller/produto/excluirProduto.php
@@ -6,17 +6,33 @@
  * ********************************************************************/
 require_once("../../config/config.php");
 require_once(SRC."/bd/produto/delete.php");
-    //testando id
-    if(deleteProduto($_GET['id'])){
+require_once(SRC."/bd/produtoCategoria/delete.php");
+    
+    $foto =(string) $_GET['foto'];
+    $preview =(string) $_GET['preview'];
+
+    if(deleteProdutoCategoria($_GET['id'])){
+        if(deleteProduto($_GET['id'])){
+        unlink(SRC.DIRETORIO_FILE.$foto);
+        unlink(SRC.DIRETORIO_FILE.$preview);
         echo("<script>
-            alert('Produto excluído com  sucesso');
-            window.location.href='../../produto.php';
-        </script>");
+                alert('Produto excluído com  sucesso');
+                window.location.href='../../produto.php';
+            </script>");
+        }
+        else{
+            echo("<script>
+                alert('Erro ao excluir Produto');
+                window.location.href='../../produto.php';
+            </script>");
+        }  
     }
     else{
         echo("<script>
-            alert('Erro ao excluir Usuário');
-            window.location.href='../../produto.php';
-        </script>");
+                alert('Erro ao excluir Produto');
+                window.location.href='../../produto.php';
+            </script>");
     }
+    //testando id
+    
 ?>
